microservices/gateway/requirements.txt
fastapi
uvicorn[standard]
gunicorn
aiostream
websockets
pyjwt
aiohttp

-e git+https://${GH_TOKEN}@github.com/lognes-delve/delve-common.git@master#egg=delve_common